Move AudioPlayer class to separate  module
